chore: circular locale types
